Added code to initialise the second display which has a GC9A01 chip.
However I don't know how to select the device I want to talk to.
You'd think it's related to spi_bus_add_device()
but I don't see that yet.

I haven't tried this and it won't work. At least there is
nothing to draw on GC9A01 and the chip is not selected.